build usecase from request
