+++ backend/api/src/presentation/http/usecases/mod.rs
@@ -1,0 +1,1 @@
+mod payment;
